refactor: split sourceAdapter, actionRegistry, and codePreview (partial #95)
Split the top 3 oversized plugin files to keep them under ~300 lines:

- sourceAdapter.ts (1062 -> 421): extracted sourceTableActions.ts,
  sourceCjkActions.ts, sourceTextTransforms.ts, sourceImageActions.ts
- actionRegistry.ts (744 -> 102): extracted actionDefinitions.ts,
  menuMapping.ts
- codePreview/tiptap.ts (738 -> 431): extracted previewHelpers.ts,
  renderers/renderLatex.ts, renderers/renderMermaidPreview.ts,
  renderers/renderMarkmapPreview.ts, renderers/renderSvgPreview.ts
